Step 3: Extract MIDI event classes into midi_events.py
Move tempo, pitch_bend, pan, volume, and event classes into new
musicpy/midi_events.py module. These classes are imported back
into structures.py to maintain backward compatibility.
